server/requirements.txt
fastapi
uvicorn[standard]
supabase
google-genai
python-dotenv
pydantic
pydantic-settings
httpx
python-multipart
